requirements.txt
GitPython>=3.1.0
requests>=2.28.0 